refactor #15 changed some imports to aliases, env changes
