--- .env
@@ -1,1 +1,0 @@
-BACKEND_API_URL=http://localhost:5172
