+++ src/app/game/components/data-menu/services/game-data-sending.service.ts
@@ -2,7 +2,7 @@ import { inject, Injectable } from '@angular/core';
 import { TExchangeData } from '../../../models/exchange-data.type';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { environment } from '../../../../../environment/environment';
+import { environment } from '@env/environment';
 
 @Injectable({
   providedIn: 'root',
